% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{cfIneq}
\alias{cfIneq}
\alias{cfIneq-package}
\title{cfIneq: Weighting Estimators of Counterfactual Distributions}
\description{
cfIneq: Weighting Estimators of Counterfactual Distributions
}
