% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{cfObj}
\alias{cfObj}
\title{cfObj}
\usage{
cfObj(ate = NULL, Df1 = NULL, Df2 = NULL, q1 = NULL, q2 = NULL,
  qte = NULL, tau = NULL, y.seq = NULL, ate.se = NULL,
  qte.se = NULL)
}
\arguments{
\item{ate}{Difference in average outcomes for 1st set of weights
relative to second set of weights}

\item{Df1}{Distribution function for first set of weights}

\item{Df2}{Distribution function for second set of weights}

\item{q1}{Quantiles for first set of weights}

\item{q2}{Quantiles for second set of weights}

\item{qte}{q1-q2}

\item{tau}{Values at which quantiles are computed}

\item{y.seq}{Values at which the distribution functions are computed;
if this is NULL, the default is to use all values of y}

\item{ate.se}{Standard error for ate}

\item{qte.se}{Standard errors for qte}
}
\value{
cfObj
}
\description{
Counterfactual objects
}
