% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{ggcf}
\alias{ggcf}
\title{ggcf}
\usage{
ggcf(cfObj, plotate = FALSE)
}
\arguments{
\item{cfObj}{A counterfactual object}

\item{plotate}{Whether or not to plot the ate too, default is FALSE}
}
\value{
ggplot object
}
\description{
Plot counterfactual quantiles using ggplot2
}
